fix: forced lowerCase transform on ansi file path (#125)
When adding a custom filepath to (.profile|.bash_rc), the previous JS
code forced a lowerCase transformation on string stored in TTC_SAY_BOX.
If the string is a path to custom ansi art, and the user follows a capitalized
directory convention on their sys, the tranformed filepath would be wrong - resulting
in an error within the tiny-care-terminal. The fix: add a simple regex to check
if the string stored in the TTC_SAY_BOX env var is indeed a filepath, if it is,
then send that string; otherwise, proceed to transform the string.