Replace underpowered git-user-name by native implementation and fix error handling
The package [`git-user-name`] merely parsed the global git configuration
without evaluating any includes therein or other configuration (system,
local).
This new implementation uses the git cli directly. The git cli can be
assumed available since it is also used by gitlog.

Further, this implements actually working error handling. If there is an
error, e.g. no git username is found or the git binary is not in
`$PATH`, then the error message is shown in the daily and weekly commits
sections.
This resolves the bug that the implementation didn't throw an error if
there was no username. Also, even if an error was thrown by the library,
the error message was not printed where the user could see it, but the
application merely exited (which is mentioned as an anti-pattern in
nodejs' [process documentation]).

[`git-user-name`]: https://github.com/jonschlinkert/git-user-name/blob/22cf1d3880356c754f7d84dec242987a3b8fcc7e/index.js
[process documentation]: https://nodejs.org/api/process.html#process_process_exit_code